+++ tests/ternary/jsfmt.spec.js
@@ -1,0 +1,9 @@
+run_spec(
+  __dirname,
+  ["apex"],
+  [
+    { useTabs: false, tabWidth: 2 },
+    { useTabs: false, tabWidth: 4 },
+    { useTabs: true, tabWidth: 2 },
+  ],
+);