+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Handle new jorje structure for types in enhanced for loops.
 - Clearer error message when failing to connect to standalone Apex parser server ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/204)).
 - Handle SOQL Select Distance Expression ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/237)).
+- Use alignment instead of indentation to provide clarity for ternary expressions ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/223)).
 
 # 1.5.0